requirements.txt
python-telegram-bot==20.7
sympy==1.12
openai==1.12.0
matplotlib==3.8.2
pillow==10.1.0
asyncpg==0.28.0  # Async PostgreSQL driver
sqlalchemy==2.0.23
alembic==1.12.1
cachetools==5.3.2
python-dotenv==1.0.1
uvicorn==0.24.0
fastapi==0.104.1
psycopg2-binary==2.9.7  # Keep for potential sync operations